Variouos refactoring to whip it into a shape.

+++ internal/execute.go
@@ -1,19 +1,24 @@
 package internal
 
 import (
+	"fmt"
 	"os"
 	"os/exec"
 )
 
 // Execute runs the generated script with the provided arguments.
-func Execute(scriptCode string, args []string) error {
+func Execute(s Script, args []string) error {
+	if s.GeneratedCode == "" {
+		return fmt.Errorf("generated script is empty")
+	}
+
 	tmpFile, err := os.CreateTemp("", "genscript-*.sh")
 	if err != nil {
 		return err
 	}
 	defer os.Remove(tmpFile.Name())
 
-	if _, err := tmpFile.WriteString(scriptCode); err != nil {
+	if _, err := tmpFile.WriteString(s.GeneratedCode); err != nil {
 		return err
 	}
 	if err := tmpFile.Close(); err != nil {
